# Make tests faster