--- a/pkg/statsd/handler_backend_test.go
+++ b/pkg/statsd/handler_backend_test.go
@@ -110,8 +110,8 @@ func TestNewBackendHandlerShouldCreateCorrectNumberOfWorkers(t *testing.T) {
 func TestRunShouldReturnWhenContextCancelled(t *testing.T) {
 	t.Parallel()
 	h := NewBackendHandler(nil, 0, 5, 1, newTestFactory())
-	ctx, cancelFunc := context.WithTimeout(context.Background(), 1*time.Second)
-	defer cancelFunc()
+	ctx, cancelFunc := context.WithCancel(context.Background())
+	cancelFunc()
 	h.Run(ctx)
 }