This started as adding a virtual clock, and then turned in to making tests fast by removing sleeps and timeouts generally.  Each commit is stand-alone.

Before (running tests 10 times, with `-short`):
```
ok      github.com/atlassian/gostatsd   0.040s
ok      github.com/atlassian/gostatsd/pkg/backends/cloudwatch   0.008s
ok      github.com/atlassian/gostatsd/pkg/backends/datadog      4.758s
ok      github.com/atlassian/gostatsd/pkg/backends/graphite     0.018s
ok      github.com/atlassian/gostatsd/pkg/backends/newrelic     4.725s
ok      github.com/atlassian/gostatsd/pkg/backends/sender       3.009s
ok      github.com/atlassian/gostatsd/pkg/backends/statsdaemon  0.009s
ok      github.com/atlassian/gostatsd/pkg/cachedinstances/cloudprovider 0.720s
ok      github.com/atlassian/gostatsd/pkg/cloudproviders/k8s    0.268s
ok      github.com/atlassian/gostatsd/pkg/cluster/nodes 0.004s
ok      github.com/atlassian/gostatsd/pkg/stats 0.106s
ok      github.com/atlassian/gostatsd/pkg/statsd        110.061s <-- the top one
ok      github.com/atlassian/gostatsd/pkg/transport     0.016s
ok      github.com/atlassian/gostatsd/pkg/web   0.521s
```

After (running tests 10 times, with `-short`):
```
ok      github.com/atlassian/gostatsd   0.048s
ok      github.com/atlassian/gostatsd/pkg/backends/cloudwatch   0.014s
ok      github.com/atlassian/gostatsd/pkg/backends/datadog      0.122s
ok      github.com/atlassian/gostatsd/pkg/backends/graphite     0.045s
ok      github.com/atlassian/gostatsd/pkg/backends/newrelic     0.121s
ok      github.com/atlassian/gostatsd/pkg/backends/sender       0.015s
ok      github.com/atlassian/gostatsd/pkg/backends/statsdaemon  0.024s
ok      github.com/atlassian/gostatsd/pkg/cachedinstances/cloudprovider 0.712s
ok      github.com/atlassian/gostatsd/pkg/cloudproviders/k8s    0.252s
ok      github.com/atlassian/gostatsd/pkg/cluster/nodes 0.017s
ok      github.com/atlassian/gostatsd/pkg/stats 0.134s
ok      github.com/atlassian/gostatsd/pkg/statsd        0.168s
ok      github.com/atlassian/gostatsd/pkg/transport     0.018s
ok      github.com/atlassian/gostatsd/pkg/web   0.521s
```

The final commit in this set (https://github.com/atlassian/gostatsd/commit/35b1a5cc79b353f24e5db47d8b26c22c012d43a2) was found by running the entire test suite thousands of times, which would have been impossible before (that particular test is one that took 11 seconds to run)